@@ -11,10 +11,18 @@
  * or submit itself to any jurisdiction.
  */
 
-const OverviewSuite = require('./overview.test');
-const DetailsSuite = require('./detail.test.js');
-
-module.exports = () => {
-    describe('Overview Page', OverviewSuite);
-    describe('Details Page', DetailsSuite);
+/**
+ * Format a percentage value
+ *
+ * @param {number|null|undefined} percentage the percentage value to format
+ * @param {number} digits the amount of digits to display
+ *
+ * @return {string} the formatted output
+ */
+export const formatPercentage = (percentage, digits = 2) => {
+    if (percentage !== undefined && percentage !== null && !isNaN(percentage)) {
+        return `${percentage.toFixed(digits)}%`;
+    } else {
+        return '-';
+    }
 };
