@@ -11,10 +11,16 @@
  * or submit itself to any jurisdiction.
  */
 
-const OverviewSuite = require('./overview.test');
-const DetailsSuite = require('./detail.test.js');
+import { lhcFillDetailsComponent } from './lhcFillDetailsComponent.js';
+import { h } from '/js/src/index.js';
 
-module.exports = () => {
-    describe('Overview Page', OverviewSuite);
-    describe('Details Page', DetailsSuite);
-};
+/**
+ * Page displaying the details of a given LHC fill
+ *
+ * @param {Model} model the global model
+ *
+ * @return {vnode} the details page
+ */
+export const LhcFillDetailsPage = (model) => h('', {
+    onremove: model.lhcFills.clearDetails,
+}, lhcFillDetailsComponent(model.lhcFills.detailsModel));
