@@ -19,11 +19,11 @@ func main() {
 	apiClient.InitializeApi(baseUrl, apiToken)
 
 	// Create a run
-	apiClient.CreateRun("go-api-Timestamp", 5, 5, 1, 80, sw.COSMICS_RunType, false, false, true, "normal",
+	apiClient.CreateRun("go-api-Timestamp", 5, 5, 1, 80, sw.COSMICS_RunType, false, false, true, "normal", "hash",
 		(sw.CPV_Detectors + "," + sw.ZDC_Detectors + "," + sw.EMC_Detectors))
 
 	// Update a run
-	apiClient.UpdateRun(80, sw.BAD_RunQuality, -1, testTime, -1, testTime, "LTU", "Repository Hash", "production/production.desc", "processing", "lhc22b")
+	apiClient.UpdateRun(80, sw.BAD_RunQuality, -1, testTime, -1, testTime, "LTU", "Repository Hash", "production/production.desc", "processing", "lhc22_b", "production/production-system.desc")
 
 	// Create an flp
 	apiClient.CreateFlp("flpTestName", "flpTestHost", 80)
