@@ -51,21 +51,22 @@ func InitializeApi(baseUrl string, apiKey string) {
  * @param triggerStart Time (UTC) when Trigger subsystem was started
  */
 func CreateRun(environmentId string, nDetectors int32, nEpns int32, nFlps int32,
-	runNumber int32, runType sw.RunType, dd_flp bool, dcs bool, epn bool, epnTopology string, detectors sw.Detectors) (sw.RunResponse, *http.Response, error) {
+	runNumber int32, runType sw.RunType, dd_flp bool, dcs bool, epn bool, epnTopology string, odcTopologyFullName string, detectors sw.Detectors) (sw.RunResponse, *http.Response, error) {
 	var run sw.RunType = runType
 	var dets sw.Detectors = detectors
 	obj := sw.Run{
-		EnvironmentId: environmentId,
-		NDetectors:    nDetectors,
-		NEpns:         nEpns,
-		NFlps:         nFlps,
-		RunNumber:     runNumber,
-		RunType:       &run,
-		DdFlp:         dd_flp,
-		Dcs:           dcs,
-		Epn:           epn,
-		EpnTopology:   epnTopology,
-		Detectors:     &dets,
+		EnvironmentId:       environmentId,
+		NDetectors:          nDetectors,
+		NEpns:               nEpns,
+		NFlps:               nFlps,
+		RunNumber:           runNumber,
+		RunType:             &run,
+		DdFlp:               dd_flp,
+		Dcs:                 dcs,
+		Epn:                 epn,
+		EpnTopology:         epnTopology,
+		Detectors:           &dets,
+		OdcTopologyFullName: odcTopologyFullName,
 	}
 
 	arrayResponse, response, err := api.RunApi.CreateRun(auth, obj)
@@ -82,7 +83,7 @@ func CreateRun(environmentId string, nDetectors int32, nEpns int32, nFlps int32,
  * @param triggerEnd (UTC) Time when Trigger subsystem was stopped
  */
 func UpdateRun(runNumber int32, runQuality sw.RunQuality, timeO2Start int64, timeO2End int64, timeTrgStart int64, timeTrgEnd int64,
-	triggerValue string, pdpConfigOption string, pdpTopologyDescriptionLibraryFile string, tfbDdMode string, lhcPeriod string) (sw.RunResponse, *http.Response, error) {
+	triggerValue string, pdpConfigOption string, pdpTopologyDescriptionLibraryFile string, tfbDdMode string, lhcPeriod string, odcTopologyFullName string) (sw.RunResponse, *http.Response, error) {
 	var runquality sw.RunQuality = runQuality
 	obj := sw.Run{
 		RunQuality:                        &runquality,
@@ -91,6 +92,7 @@ func UpdateRun(runNumber int32, runQuality sw.RunQuality, timeO2Start int64, tim
 		TfbDdMode:                         tfbDdMode,
 		LhcPeriod:                         lhcPeriod,
 		TriggerValue:                      triggerValue,
+		OdcTopologyFullName:               odcTopologyFullName,
 	}
 	if timeO2End != -1 {
 		obj.TimeO2End = timeO2End
