@@ -62,7 +62,7 @@ module.exports = {
         { text: 'TEST-TAG-40' },
         { text: 'TEST-TAG-41' },
         { text: 'TEST-TAG-42' },
-        { text: 'TEST-TAG-43' },
+        { text: 'PHYSICS' },
     ]),
 
     down: (queryInterface, _Sequelize) => queryInterface.bulkDelete('tags', null, {}),
