@@ -36,6 +36,7 @@ module.exports = {
             detectors: 'CPV',
             trigger_value: 'OFF',
             lhc_period: 'lhc22b',
+            odc_topology_full_name: 'hash',
         },
         {
             id: 2,
@@ -59,6 +60,7 @@ module.exports = {
             detectors: 'CPV',
             trigger_value: 'OFF',
             lhc_period: 'lhc22b',
+            odc_topology_full_name: 'hash',
         },
         {
             id: 3,
@@ -173,6 +175,7 @@ module.exports = {
             epn_topology: 'normal',
             detectors: 'CPV',
             trigger_value: 'LTU',
+            odc_topology_full_name: 'hash',
         },
         {
             id: 8,
@@ -196,6 +199,7 @@ module.exports = {
             epn_topology: 'normal',
             detectors: 'CPV',
             trigger_value: 'LTU',
+            odc_topology_full_name: 'production/production.desc',
         },
         {
             id: 9,
@@ -219,6 +223,7 @@ module.exports = {
             epn_topology: 'normal',
             detectors: 'CPV',
             trigger_value: 'CTP',
+            odc_topology_full_name: 'hash',
         },
         {
             id: 10,
@@ -241,6 +246,7 @@ module.exports = {
             epn_topology: 'normal',
             detectors: 'CPV',
             trigger_value: 'CTP',
+            odc_topology_full_name: 'production/production.desc',
         },
         {
             id: 11,
@@ -263,6 +269,7 @@ module.exports = {
             epn_topology: 'normal',
             detectors: 'CPV',
             trigger_value: 'CTP',
+            odc_topology_full_name: 'production/production.desc',
         },
         {
             id: 12,
@@ -285,6 +292,7 @@ module.exports = {
             epn_topology: 'normal',
             detectors: 'CPV',
             trigger_value: 'CTP',
+            odc_topology_full_name: 'production/production.desc',
         },
         {
             id: 13,
@@ -1183,23 +1191,24 @@ module.exports = {
             id: 54,
             run_number: 54,
             time_o2_start: '2019-08-08 13:00:00',
-            time_o2_end: '2019-08-08 13:00:00',
-            time_trg_start: '2019-08-08 13:00:00',
-            time_trg_end: '2019-08-08 13:00:00',
+            time_o2_end: '2019-08-08 17:00:00',
+            time_trg_start: '2019-08-08 14:00:00',
+            time_trg_end: '2019-08-08 16:00:00',
             environment_id: '1234567890',
             run_type: 'physics',
-            run_quality: 'test',
-            n_detectors: 6,
+            run_quality: 'good',
+            n_detectors: 2,
             n_flps: 12,
             n_epns: 12,
             n_subtimeframes: 12,
             bytes_read_out: 64,
             dd_flp: true,
-            dcs: false,
+            dcs: true,
             epn: true,
             epn_topology: 'normal',
-            detectors: 'CPV',
-            fill_number: 5,
+            detectors: 'ITS',
+            trigger_value: 'CTP',
+            fill_number: 6,
             lhc_beam_energy: 23.21,
             lhc_beam_mode: 'STABLE',
             lhc_beta_star: 321.1321,
@@ -1210,44 +1219,48 @@ module.exports = {
         {
             id: 55,
             run_number: 55,
-            time_o2_start: '2019-08-08 13:00:00',
-            time_o2_end: '2019-08-08 13:00:00',
-            time_trg_start: '2019-08-08 13:00:00',
-            time_trg_end: '2019-08-08 13:00:00',
+            time_o2_start: '2019-08-08 14:00:00',
+            time_o2_end: '2019-08-08 19:00:00',
+            time_trg_start: '2019-08-08 16:00:00',
+            time_trg_end: '2019-08-08 18:00:00',
             environment_id: '1234567890',
-            run_type: 'cosmics',
-            run_quality: 'test',
-            n_detectors: 3,
+            run_type: 'physics',
+            run_quality: 'good',
+            n_detectors: 1,
             n_flps: 12,
             n_epns: 12,
             n_subtimeframes: 12,
             bytes_read_out: 64,
             dd_flp: true,
-            dcs: false,
+            dcs: true,
             epn: true,
             epn_topology: 'normal',
-            detectors: 'CPV',
+            detectors: 'FT0',
+            trigger_value: 'CTP',
+            fill_number: 6,
         },
         {
             id: 56,
             run_number: 56,
-            time_o2_start: '2019-08-08 13:00:00',
-            time_o2_end: '2019-08-08 13:00:00',
-            time_trg_start: '2019-08-08 13:00:00',
-            time_trg_end: '2019-08-08 13:00:00',
+            time_o2_start: '2019-08-08 19:00:00',
+            time_o2_end: '2019-08-08 22:00:00',
+            time_trg_start: '2019-08-08 20:00:00',
+            time_trg_end: '2019-08-08 21:00:00',
             environment_id: '1234567890',
-            run_type: 'cosmics',
+            run_type: 'physics',
             run_quality: 'good',
-            n_detectors: 6,
+            n_detectors: 2,
             n_flps: 12,
             n_epns: 12,
             n_subtimeframes: 12,
             bytes_read_out: 64,
             dd_flp: true,
-            dcs: false,
+            dcs: true,
             epn: true,
             epn_topology: 'normal',
-            detectors: 'CPV',
+            detectors: 'ITS, FT0',
+            trigger_value: 'CTP',
+            fill_number: 6,
         },
         {
             id: 57,
@@ -2299,7 +2312,7 @@ module.exports = {
             environment_id: '0987654321',
             run_type: 'physics',
             run_quality: 'good',
-            n_detectors: 3,
+            n_detectors: 15,
             n_flps: 12,
             n_epns: 12,
             n_subtimeframes: 12,
@@ -2311,6 +2324,7 @@ module.exports = {
             epn_topology: 'a quite long topology, which will for sure require a balloon to be displayed properly',
             detectors: 'ACO, CPV, CTP, EMC, FIT, HMP, ITS, MCH, MFT, MID, PHS, TOF, TPC, TRD, ZDC',
             lhc_period: 'lhc22b',
+            odc_topology_full_name: 'hash',
         },
     ]),
 
