@@ -45,9 +45,6 @@ module.exports = {
         },
         {
             fill_number: 4,
-            stable_beams_start: '2022-03-21 13:00:00',
-            stable_beams_end: '2022-03-22 15:00:00',
-            stable_beams_duration: 100,
             beam_type: 'p-p',
             filling_scheme_name: 'schemename',
             created_at: new Date('2019-08-09'),
@@ -56,13 +53,19 @@ module.exports = {
         {
             fill_number: 5,
             stable_beams_start: '2022-03-21 13:00:00',
-            stable_beams_end: '2022-03-22 15:00:00',
-            stable_beams_duration: 100,
+            filling_scheme_name: 'schemename',
+            created_at: new Date('2019-08-09'),
+            updated_at: new Date('2019-08-09 16:00:00'),
+        },
+        {
+            fill_number: 6,
+            stable_beams_start: '2019-08-08 11:00:00',
+            stable_beams_end: '2019-08-08 23:00:00',
+            stable_beams_duration: 60 * 60 * 12,
             filling_scheme_name: 'schemename',
             created_at: new Date('2019-08-09'),
             updated_at: new Date('2019-08-09 16:00:00'),
         },
     ]),
     down: (queryInterface, _Sequelize) => queryInterface.bulkDelete('lhc_fills', null, {}),
-
 };
