@@ -28,7 +28,11 @@ module.exports = {
         },
         {
             run_id: 106,
-            tag_id: 7,
+            tag_id: 8,
+        },
+        {
+            run_id: 2,
+            tag_id: 50,
         },
     ]),
 
