@@ -57,7 +57,7 @@ const FilterDto = Joi.object({
     ddflp: Joi.boolean(),
     dcs: Joi.boolean(),
     epn: Joi.boolean(),
-    epnTopology: Joi.string().trim(),
+    odcTopologyFullName: Joi.string().trim(),
     detectors: Joi.string().trim(),
     lhcPeriods: Joi.string().trim(),
 });
