@@ -118,8 +118,11 @@ export default class Model extends Observable {
             case 'env-overview':
                 this.envs.fetchAllEnvs();
                 break;
-            case 'lhcFill-overview':
+            case 'lhc-fill-overview':
                 this.lhcFills.fetchAllLhcFills();
+                break;
+            case 'lhc-fill-details':
+                this.lhcFills.loadDetails(this.router.params.fillNumber);
                 break;
             case 'log-overview':
                 // Prevent loading extra rows when it's not necessary
