@@ -14,23 +14,17 @@
 import { h } from '/js/src/index.js';
 
 /**
- * Generic error alert.
- *
- * @param {Object[]} errors The JSON:API conform error object.
- * @returns {vnode[]} The list of vnodes representing the alerts.
+ * Returns the title filter component
+ * @param {Object} model The global model object
+ * @return {vnode} A text box that allows the user to enter a title substring to match against all logs
  */
-const errorAlert = (errors) => {
-    if (!Array.isArray(errors)) {
-        errors = [errors];
-    }
+const odcTopologyFullName = (model) => h('input.w-75.mt1', {
+    type: 'text',
+    id: 'Topology Full Name',
+    value: model.runs.odcTopologyFullNameFilter,
+    oninput: (e) => {
+        model.runs.odcTopologyFullNameFilter = e.target.value;
+    },
+}, '');
 
-    return errors.map((error) => {
-        if (error.detail) {
-            return h('.alert.alert-danger', h('b', `${error.title}: `), error.detail);
-        } else {
-            return h('.alert.alert-danger', h('b', error.title));
-        }
-    });
-};
-
-export default errorAlert;
+export default odcTopologyFullName;
