@@ -60,7 +60,7 @@ const navBar = (model) => {
             pageTab(model, 'home', 'Home'),
             pageTab(model, 'log-overview', 'Log Entries'),
             pageTab(model, 'env-overview', 'Environments'),
-            pageTab(model, 'lhcFill-overview', 'LHC Fills'),
+            pageTab(model, 'lhc-fill-overview', 'LHC Fills'),
             pageTab(model, 'run-overview', 'Runs'),
             h(`.navbar-dropdown${model.dropdownMenu === 'overview' ? '.dropdown-open' : ''}`, [
                 h('button#overviews', {
