@@ -11,7 +11,7 @@
  * or submit itself to any jurisdiction.
  */
 
-import format from '../Detail/index.js';
+import { detailsList } from '../Detail/detailsList.js';
 import { h, iconTrash, iconPlus } from '/js/src/index.js';
 import { formatTimestamp } from '../../utilities/formatting/formatTimestamp.js';
 import { formatBoolean } from '../../utilities/formatting/formatBoolean.js';
@@ -80,16 +80,15 @@ const activeFields = (model) => ({
                     model.runs.runChanges = { key: 'runQuality', value: target.value };
                 },
             }, [
-                h('option#runQualityGood', {
-                    value: 'good', selected: model.runs.run.payload.runQuality === 'good' }, 'good'),
-                h('option#runQualityBad', {
-                    value: 'bad', selected: model.runs.run.payload.runQuality === 'bad' }, 'bad'),
-                h('option#runQualityTest', {
-                    value: 'test', selected: model.runs.run.payload.runQuality === 'test' }, 'test'),
+                h('option#runQualityGood', { value: 'good', selected: model.runs.run.payload.runQuality === 'good' }, 'good'),
+                h('option#runQualityBad', { value: 'bad', selected: model.runs.run.payload.runQuality === 'bad' }, 'bad'),
+                h('option#runQualityTest', { value: 'test', selected: model.runs.run.payload.runQuality === 'test' }, 'test'),
             ]),
-        ] : runQuality ? h('.badge.white', {
-            class: runQuality === 'good' ? 'bg-success' : runQuality === 'bad' ? 'bg-danger' : 'bg-gray-darker',
-        }, runQuality) : '-',
+        ] : runQuality ? h(
+            '.badge.white',
+            { class: runQuality === 'good' ? 'bg-success' : runQuality === 'bad' ? 'bg-danger' : 'bg-gray-darker' },
+            runQuality,
+        ) : '-',
     },
     nDetectors: {
         name: 'Number of Detectors',
@@ -122,24 +121,6 @@ const activeFields = (model) => ({
         name: 'TFB DD Mode',
         visible: true,
     },
-
-    /*
-     * NEpns: {
-     *     name: 'Number of Epns',
-     *     visible: false,
-     *     size: 'cell-s',
-     * },
-     * nSubtimeframes: {
-     *     name: 'Number of STFs',
-     *     visible: false,
-     *     size: 'cell-s',
-     * },
-     * bytesReadOut: {
-     *     name: 'Readout Data',
-     *     visible: false,
-     *     size: 'cell-m',
-     * },
-     */
     dd_flp: {
         name: 'Data Distribution (FLP)',
         visible: true,
@@ -159,7 +140,12 @@ const activeFields = (model) => ({
         format: formatBoolean,
     },
     epnTopology: {
-        name: 'EPN Topology',
+        name: 'Topology',
+        visible: true,
+        size: 'cell-m',
+    },
+    odcTopologyFullName: {
+        name: 'Topology Full Name',
         visible: true,
         size: 'cell-m',
     },
@@ -180,7 +166,7 @@ const activeFields = (model) => ({
  */
 const entry = (model, post) => {
     const postFields = activeFields(model);
-    return format(postFields, post, 'Run', '-');
+    return detailsList(postFields, post, 'Run', '-');
 };
 
 /**
