@@ -11,7 +11,7 @@
  * or submit itself to any jurisdiction.
  */
 
-import format from '../Detail/index.js';
+import { detailsList } from '../Detail/detailsList.js';
 import { h } from '/js/src/index.js';
 import { formatTimestamp } from '../../utilities/formatting/formatTimestamp.js';
 
@@ -82,7 +82,7 @@ const activeFields = (model) => ({
  */
 const entry = (model, post) => {
     const postFields = activeFields(model);
-    return format(postFields, post, 'Tag');
+    return detailsList(postFields, post, 'Tag');
 };
 
 export default entry;
