@@ -11,18 +11,14 @@
  * or submit itself to any jurisdiction.
  */
 
-import { h } from '/js/src/index.js';
-
 /**
- * Returns the title filter component
- * @param {Object} model The global model object
- * @return {vnode} A text box that allows the user to enter a title substring to match against all logs
+ * States if the given run may be a physics run or not, depending on if it is related to a stable beam
+ *
+ * @param {object} run the run to check for definition
+ *
+ * @return {boolean} true if the run is a physic run
  */
-const titleFilter = (model) => h('input.w-75.mt1', {
-    type: 'text',
-    id: 'epnTopology',
-    value: model.runs.getEpnTopologyFilter(),
-    oninput: (e) => model.runs.setEpnTopologyFilter(e.target.value),
-}, '');
-
-export default titleFilter;
+export const isPotentiallyPhysicsRun = ({ detectors, dd_flp, dcs, epn, triggerValue }) =>
+    (detectors.includes('ITS') || detectors.includes('FT0'))
+    && dd_flp && dcs && epn
+    && triggerValue === 'CTP';
