@@ -20,29 +20,14 @@
  * @returns {String} The new target URL.
  */
 const targetURL = (model, targetKey, targetValue) => {
-    let target = '';
-    let seenTarget = false;
-    for (const pair of Object.entries(model.router.params)) {
-        const [key] = pair;
-        let [, value] = pair;
+    const newParams = {
+        ...model.router.params,
+        [targetKey]: targetValue,
+    };
 
-        if (key === targetKey) {
-            value = targetValue;
-            seenTarget = true;
-        }
-
-        if (target === '') {
-            target += `?${key}=${value}`;
-        } else {
-            target += `&${key}=${value}`;
-        }
-    }
-
-    if (!seenTarget) {
-        target += `&${targetKey}=${targetValue}`;
-    }
-
-    return target;
+    // eslint-disable-next-line require-jsdoc
+    const getParameterExpression = (key, value) => `${key}=${value}`;
+    return `?${Object.entries(newParams).map(([key, value]) => getParameterExpression(key, value)).join('&')}`;
 };
 
 export default targetURL;
