@@ -30,6 +30,7 @@ import HomeOverview from './views/Home/Overview/index.js';
 import EnvOverview from './views/Envs/Overview/index.js';
 import LhcFillOverview from './views/LhcFills/Overview/index.js';
 import { modalContainer } from './components/modal/modalContainer.js';
+import { LhcFillDetailsPage } from './views/LhcFills/Detail/LhcFillDetailsPage.js';
 
 /**
  * Main view layout
@@ -46,7 +47,8 @@ export default (model) => {
 
         'env-overview': EnvOverview,
 
-        'lhcFill-overview': LhcFillOverview,
+        'lhc-fill-overview': LhcFillOverview,
+        'lhc-fill-details': LhcFillDetailsPage,
 
         'run-overview': RunsOverview,
         'run-detail': RunDetail,
