@@ -21,7 +21,7 @@ import table from '../../../components/Table/index.js';
  */
 const aboutOverview = () => {
     const data = [
-        { type: 'Bookkeeping', version: '0.35.0', hostname: '-', port: '4000' },
+        { type: 'Bookkeeping', version: '0.36.0', hostname: '-', port: '4000' },
         { type: 'NPM', version: '', hostname: '', port: '' },
     ];
 
