@@ -15,16 +15,18 @@
 import { formatTimestamp } from '../../../utilities/formatting/formatTimestamp.js';
 import { formatDuration } from '../../../utilities/formatting/formatDuration.js';
 import { frontLink } from '../../../utilities/frontLink.js';
+import { formatPercentage } from '../../../utilities/formatting/formatPercentage.js';
 
 /**
  * List of active columns for a lhc fills table
  */
 export const lhcFillsActiveColumns = {
     fillNumber: {
-        name: 'Fill Number',
+        name: 'Fill #',
         visible: true,
         primary: true,
         size: 'w-10',
+        format: (fillNumber) => frontLink(fillNumber, 'lhc-fill-details', { fillNumber: fillNumber }),
     },
     stableBeamsStart: {
         name: 'Stable beams start',
@@ -49,6 +51,30 @@ export const lhcFillsActiveColumns = {
         visible: true,
         size: 'w-15',
         format: (value) => value ? value : '-',
+    },
+    efficiency: {
+        name: 'Fill Efficiency',
+        visible: true,
+        size: 'w-15',
+        format: (efficiency) => formatPercentage(efficiency),
+    },
+    durationBeforeFirstRun: {
+        name: 'Before 1st run',
+        visible: true,
+        size: 'w-15',
+        format: (duration, lhcFill) => duration ? `${formatDuration(duration)} (${formatPercentage(lhcFill.efficiencyLossAtStart)})` : '-',
+    },
+    meanRunDuration: {
+        name: 'Mean run duration',
+        visible: true,
+        size: 'w-15',
+        format: (duration) => formatDuration(duration),
+    },
+    totalRunsDuration: {
+        name: 'Total runs duration',
+        visible: true,
+        size: 'w-15',
+        format: (duration) => formatDuration(duration),
     },
     fillingSchemeName: {
         name: 'Scheme name',
