@@ -12,11 +12,13 @@
  */
 
 import { fetchClient, Observable, RemoteData } from '/js/src/index.js';
+import { LhcFillDetailsModel } from './Detail/LhcFillDetailsModel.js';
+import { addStatisticsToLhcFill } from '../../services/lhcFill/addStatisticsToLhcFill.js';
 
 /**
  * Model representing handlers for LhcFills (lhcFills)
  */
-export default class Overview extends Observable {
+export default class LhcFills extends Observable {
     /**
      * The constructor of the Overview model object
      * @param {Object} model Pass the model to access the defined functions
@@ -25,11 +27,34 @@ export default class Overview extends Observable {
     constructor(model) {
         super();
         this.model = model;
+
+        // Sub-models
+        this._detailsModel = new LhcFillDetailsModel();
+        this._detailsModel.bubbleTo(this);
+
         this._lhcFillsPerPage = 10;
         this.clearLhcFills();
         this.resetLhcFillsParams(false);
+    }
 
-        this._createdLhcFill = RemoteData.NotAsked();
+    /**
+     * Load a new LHC fill details model for the given fill id
+     *
+     * @param {number} lhcFillNumber the number of the fill for which model details must be instantiated
+     *
+     * @return {void}
+     */
+    loadDetails(lhcFillNumber) {
+        this._detailsModel.lhcFillNumber = lhcFillNumber;
+    }
+
+    /**
+     * Clear the details model
+     *
+     * @return {void}
+     */
+    clearDetails() {
+        this._detailsModel.lhcFillNumber = null;
     }
 
     /**
@@ -51,6 +76,11 @@ export default class Overview extends Observable {
         const result = await response.json();
 
         if (result.data) {
+            // Add statistics to runs
+            result.data.forEach((lhcFill) => {
+                addStatisticsToLhcFill(lhcFill);
+            });
+
             if (this.isInfiniteScrollEnabled()) {
                 const payload = this._lhcFills && savedValues.payload ? [...savedValues.payload, ...result.data] : result.data;
                 this._lhcFills = RemoteData.success(payload);
@@ -68,41 +98,6 @@ export default class Overview extends Observable {
             ]);
         }
         this.notify();
-    }
-
-    /**
-     * Retrieve a specified log from the API
-     * @param {Number} id The ID of the log to be found
-     * @returns {undefined} Injects the data object with the response data
-     */
-    async fetchOneLhcFill(id) {
-        this._lhcFills = RemoteData.loading();
-        this.notify();
-
-        const response = await fetchClient(`/api/lhcFill/${id}`, { method: 'GET' });
-        const result = await response.json();
-
-        if (result.data) {
-            this._lhcFill = RemoteData.success([result.data]);
-            this._lhcFill = this._lhcFill.payload.tags.map(({ id }) => id);
-        } else {
-            this._lhcFill = RemoteData.failure(result.errors || [
-                {
-                    title: result.error,
-                    detail: result.message,
-                },
-            ]);
-        }
-        this.notify();
-    }
-
-    /**
-     * Clears the chosen lhcFillironment.
-     * @returns {undefined}
-     */
-    clearLhcFill() {
-        this._lhcFill = RemoteData.NotAsked();
-        this._lhcFillRuns = RemoteData.NotAsked();
     }
 
     /**
@@ -295,5 +290,14 @@ export default class Overview extends Observable {
      */
     getSelectedPage() {
         return this._selectedPage;
+    }
+
+    /**
+     * Returns the model for the details page
+     *
+     * @return {LhcFillDetailsModel} the details model
+     */
+    get detailsModel() {
+        return this._detailsModel;
     }
 }
