@@ -18,7 +18,7 @@ import o2endFilter from '../../../components/Filters/RunsFilter/o2stop.js';
 import environmentIdFilter from '../../../components/Filters/RunsFilter/environmentId.js';
 import nDetectorsFilter from '../../../components/Filters/RunsFilter/nDetectors.js';
 import nFlpsFilter from '../../../components/Filters/RunsFilter/nFlps.js';
-import epnTopologyFilter from '../../../components/Filters/RunsFilter/epnTopology.js';
+import odcTopologyFullName from '../../../components/Filters/RunsFilter/odcTopologyFullName.js';
 import detectorsFilter from '../../../components/Filters/RunsFilter/detectors.js';
 import ddflpFilter from '../../../components/Filters/RunsFilter/ddflp.js';
 import dcsFilter from '../../../components/Filters/RunsFilter/dcs.js';
@@ -29,6 +29,7 @@ import { formatTimestamp } from '../../../utilities/formatting/formatTimestamp.j
 import { durationFilter } from '../../../components/Filters/RunsFilter/durationFilter.js';
 import { formatRunDuration } from '../formatRunDuration.js';
 import fillNumbersFilter from '../../../components/Filters/RunsFilter/fillNumbers.js';
+import { frontLink } from '../../../utilities/frontLink.js';
 import nEpnsFilter from '../../../components/Filters/RunsFilter/nEpns.js';
 import triggerValueFilter from '../../../components/Filters/RunsFilter/triggerValue.js';
 import lhcPeriodsFilter from '../../../components/Filters/RunsFilter/lhcPeriod.js';
@@ -75,7 +76,7 @@ export const runsActiveColumns = {
         name: 'Fill No.',
         visible: true,
         size: 'w-5 f6',
-        format: (fill) => fill !== null ? fill : '-',
+        format: (fill) => fill !== null ? frontLink(fill, 'lhc-fill-details', { fillNumber: fill }) : '-',
         filter: fillNumbersFilter,
     },
     lhcPeriod: {
@@ -203,9 +204,15 @@ export const runsActiveColumns = {
     },
     epnTopology: {
         name: 'EPN Topology',
-        visible: true,
+        visible: false,
         size: 'w-15 f6',
-        filter: epnTopologyFilter,
+        balloon: true,
+    },
+    odcTopologyFullName: {
+        name: 'Topology Full',
+        size: 'w-15 f6',
+        visible: true,
+        filter: odcTopologyFullName,
         balloon: true,
     },
 };
