@@ -14,22 +14,26 @@
 
 import errorAlert from '../../../components/common/errorAlert.js';
 import spinner from '../../../components/common/spinner.js';
-import lhcFillDetail from '../../../components/lhcFillsDetail/index.js';
+import { h } from '/js/src/index.js';
+import { lhcFillDisplayConfiguration } from '../../LhcFills/Detail/lhcFillDisplayConfiguration.js';
+import { detailsList } from '../../../components/Detail/detailsList.js';
+import { runLhcFillDisplayConfiguration } from './runLhcFillDisplayConfiguration.js';
 
 /**
  * Show the lhc data panel
  * @param {Object} model General model
- * @returns {vNode} Value of run
+ * @returns {vnode} Value of run
  */
 const lhcFillPanel = (model) => {
     const { lhcFill } = model.runs;
     return lhcFill.match({
         NotAsked: () => model.runs.fetchLhcFill(),
         Loading: () => spinner({ size: 5, absolute: false }),
-        Success: (value) => lhcFillDetail(model, {
-            ...value,
-            ...model.runs.getRun().payload,
-        }),
+        Success: (payload) => [
+            h('strong', 'LHC Data'),
+            detailsList(lhcFillDisplayConfiguration, payload, 'lhc-fill', '-'),
+            detailsList(runLhcFillDisplayConfiguration, model.runs.getRun().payload, 'run-lhc-fill', '-'),
+        ],
         Failure: (error) => error.map(errorAlert),
     });
 };
