@@ -408,7 +408,7 @@ export default class RunModel extends Observable {
 
         this.epnFilter = '';
 
-        this.epnTopologyFilterValues = '';
+        this._odcTopologyFullNameFilter = '';
 
         this.detectorsFilterValues = '';
 
@@ -456,7 +456,7 @@ export default class RunModel extends Observable {
             || this.ddflpFilter !== ''
             || this.dcsFilter !== ''
             || this.epnFilter !== ''
-            || this.epnTopologyFilterValues !== ''
+            || this._odcTopologyFullNameFilter !== ''
             || this.detectorsFilterValues !== ''
         );
     }
@@ -522,8 +522,8 @@ export default class RunModel extends Observable {
         if (this.epnFilter !== '') {
             this.activeFilters.push('EPN');
         }
-        if (this.epnTopologyFilterValues !== '') {
-            this.activeFilters.push('Epn topology');
+        if (this._odcTopologyFullNameFilter !== '') {
+            this.activeFilters.push('Topology');
         }
         if (this.detectorsFilterValues !== '') {
             this.activeFilters.push('Detectors');
@@ -1059,8 +1059,8 @@ export default class RunModel extends Observable {
      * Returns the current epnTopology substring filter
      * @returns {String} The current epnTopology substring filter
      */
-    getEpnTopologyFilter() {
-        return this.epnTopologyFilter;
+    get odcTopologyFullNameFilter() {
+        return this._odcTopologyFullNameFilter;
     }
 
     /**
@@ -1068,8 +1068,8 @@ export default class RunModel extends Observable {
      * @param {String} newTopology The epnTopology substring to apply to the filter
      * @returns {undefined}
      */
-    setEpnTopologyFilter(newTopology) {
-        this.epnTopologyFilter = newTopology.trim();
+    set odcTopologyFullNameFilter(newTopology) {
+        this._odcTopologyFullNameFilter = newTopology.trim();
         this._applyFilters();
     }
 
@@ -1370,8 +1370,8 @@ export default class RunModel extends Observable {
             ...(this.epnFilter === true || this.epnFilter === false) && {
                 'filter[epn]': this.epnFilter,
             },
-            ...this.epnTopologyFilter && {
-                'filter[epnTopology]': this.epnTopologyFilter,
+            ...this._odcTopologyFullNameFilter && {
+                'filter[odcTopologyFullName]': this._odcTopologyFullNameFilter,
             },
             ...this.detectorsFilterValues && {
                 'filter[detectors]': this.detectorsFilterValues,
