@@ -39,11 +39,9 @@ const { dtoValidator } = require('../utilities');
  *                         string, parameters, body, HTTP headers, and so on.
  * @param {Object} response The *response* object represents the HTTP response that an Express app sends when it gets an
  *                          HTTP request.
- * @param {Function} next The *next* object represents the next middleware function which is used to pass control to the
- *                        next middleware function.
  * @returns {undefined}
  */
-const createLhcFill = async (request, response, next) => {
+const createLhcFill = async (request, response) => {
     const value = await dtoValidator(CreateLhcFillDto, request, response);
 
     if (!value) {
@@ -67,11 +65,9 @@ const createLhcFill = async (request, response, next) => {
  *                         string, parameters, body, HTTP headers, and so on.
  * @param {Object} response The *response* object represents the HTTP response that an Express app sends when it gets an
  *                          HTTP request.
- * @param {Function} next The *next* object represents the next middleware function which is used to pass control to the
- *                        next middleware function.
  * @returns {undefined}
  */
-const listLhcFills = async (request, response, next) => {
+const listLhcFills = async (request, response) => {
     const value = await dtoValidator(GetAllLhcFillsDto, request, response);
     if (!value) {
         return;
@@ -100,11 +96,9 @@ const listLhcFills = async (request, response, next) => {
  *                         string, parameters, body, HTTP headers, and so on.
  * @param {Object} response The *response* object represents the HTTP response that an Express app sends when it gets an
  *                          HTTP request.
- * @param {Function} next The *next* object represents the next middleware function which is used to pass control to the
- *                        next middleware function.
  * @returns {undefined}
  */
-const updateLhcFills = async (request, response, next) => {
+const updateLhcFills = async (request, response) => {
     const value = await dtoValidator(UpdateLhcFillDto, request, response);
     if (!value) {
         return;
@@ -127,11 +121,9 @@ const updateLhcFills = async (request, response, next) => {
  *                         string, parameters, body, HTTP headers, and so on.
  * @param {Object} response The *response* object represents the HTTP response that an Express app sends when it gets an
  *                          HTTP request.
- * @param {Function} next The *next* object represents the next middleware function which is used to pass control to the
- *                        next middleware function.
  * @returns {undefined}
  */
-const getLhcFillById = async (request, response, next) => {
+const getLhcFillById = async (request, response) => {
     const value = await dtoValidator(GetLhcFillDto, request, response);
 
     if (!value) {
@@ -164,11 +156,9 @@ const getLhcFillById = async (request, response, next) => {
  *                         string, parameters, body, HTTP headers, and so on.
  * @param {Object} response The *response* object represents the HTTP response that an Express app sends when it gets an
  *                          HTTP request.
- * @param {Function} next The *next* object represents the next middleware function which is used to pass control to the
- *                        next middleware function.
  * @returns {undefined}
  */
-const listRuns = async (request, response, next) => {
+const listRuns = async (request, response) => {
     const value = await dtoValidator(GetLhcFillDto, request, response);
     if (!value) {
         return;
@@ -186,11 +176,9 @@ const listRuns = async (request, response, next) => {
  *                         string, parameters, body, HTTP headers, and so on.
  * @param {Object} response The *response* object represents the HTTP response that an Express app sends when it gets an
  *                          HTTP request.
- * @param {Function} next The *next* object represents the next middleware function which is used to pass control to the
- *                        next middleware function.
  * @returns {undefined}
  */
-const getRunByRunNumber = async (request, response, next) => {
+const getRunByRunNumber = async (request, response) => {
     const value = await dtoValidator(GetLhcFillRunDto, request, response);
     if (!value) {
         return;
