@@ -25,6 +25,7 @@ const {
         QueryBuilder,
         TransactionHelper,
     },
+    models,
 } = require('../../database');
 
 /**
@@ -42,7 +43,7 @@ class GetLhcFillUseCase {
         const { fillNumber } = params;
 
         const queryBuilder = new QueryBuilder()
-            .include('runs')
+            .include({ model: models.Run, as: 'runs', include: 'tags' })
             .where('fillNumber').is(fillNumber);
 
         const lhcFill = await TransactionHelper.provide(() => LhcFillRepository.findOne(queryBuilder));
