@@ -58,7 +58,7 @@ class GetAllRunsUseCase {
                 ddflp,
                 dcs,
                 epn,
-                epnTopology,
+                odcTopologyFullName,
                 detectors,
                 lhcPeriods,
             } = filter;
@@ -147,8 +147,8 @@ class GetAllRunsUseCase {
             } else if (epn === true) {
                 queryBuilder.where('epn').is(true);
             }
-            if (epnTopology) {
-                queryBuilder.where('epnTopology').substring(epnTopology);
+            if (odcTopologyFullName) {
+                queryBuilder.where('odcTopologyFullName').substring(odcTopologyFullName);
             }
             if (detectors) {
                 queryBuilder.where('detectors').substring(detectors);
