@@ -60,9 +60,9 @@ module.exports = () => {
 
     it('should show lhc data in edit mode', async () => {
         await page.waitForTimeout(100);
-        const element = await page.$('#lhcFill-id>b');
+        const element = await page.$('#lhc-fill-fillNumber>strong');
         const value = await element.evaluate((el) => el.textContent);
-        expect(value).to.equal('LHC Data:');
+        expect(value).to.equal('Fill number:');
     });
 
     it('successfully exited EDIT mode of a run', async () => {
@@ -86,9 +86,9 @@ module.exports = () => {
     });
     it('should show lhc data in normal mode', async () => {
         await page.waitForTimeout(100);
-        const element = await page.$('#lhcFill-id>b');
+        const element = await page.$('#lhc-fill-fillNumber>strong');
         const value = await element.evaluate((el) => el.textContent);
-        expect(value).to.equal('LHC Data:');
+        expect(value).to.equal('Fill number:');
     });
     it('can navigate to a log detail page', async () => {
         table = await page.$$('tr');
@@ -99,6 +99,25 @@ module.exports = () => {
         await page.waitForTimeout(300);
         const redirectedUrl = await page.url();
         expect(String(redirectedUrl).startsWith(`${url}/?page=log-detail&id=1`)).to.be.true;
+    });
+
+    it('should successfully navigate to the LHC fill details page', async () => {
+        await page.goto(`${url}/?page=run-detail&id=106`, { waitUntil: 'networkidle0' });
+        await page.waitForTimeout(100);
+
+        const fillNumberSelector = '#lhc-fill-fillNumber a';
+        // Remove "row" prefix to get fill number
+        const fillNumber = await page.$eval(fillNumberSelector, (element) => element.innerText);
+
+        await page.$eval(fillNumberSelector, (link) => link.click());
+        await page.waitForNetworkIdle();
+        await page.waitForTimeout(100);
+
+        const redirectedUrl = await page.url();
+        const urlParameters = redirectedUrl.slice(redirectedUrl.indexOf('?') + 1).split('&');
+
+        expect(urlParameters).to.contain('page=lhc-fill-details');
+        expect(urlParameters).to.contain(`fillNumber=${fillNumber}`);
     });
 
     it('notifies if a specified run id is invalid', async () => {
