@@ -100,7 +100,7 @@ module.exports = () => {
         await page.type(inputSelector, '-TAG-');
         await page.waitForTimeout(300);
         let table = await page.$$('tbody tr');
-        expect(table.length).to.equal(43);
+        expect(table.length).to.equal(42);
 
         await page.type(inputSelector, 'DO-NOT-EXIST');
         await page.waitForTimeout(200);
