@@ -88,7 +88,7 @@ module.exports = () => {
         ({ tags } = await new GetAllTagsUseCase().execute(getAllTagsDto));
 
         expect(tags).to.be.an('array');
-        expect(tags).to.have.lengthOf(43);
+        expect(tags).to.have.lengthOf(42);
         expect(tags.every((tag) => tag.text.includes('-TAG-'))).to.be.true;
 
         getAllTagsDto.query = { filter: { partialText: 'DO-NOT-EXISTS' } };
